+++ tests/test_editor.py
@@ -44,6 +44,7 @@ def test_make_editor_session(pad, path, kwargs, expect):
             marks=pytest.mark.xfail(reason="buglet that should be fixed"),
         ),
         ("/../../templates", {}, "Invalid path"),
+        ("/..\\..\\templates", {}, "Invalid path"),
     ],
 )
 def test_make_editor_session_raises_bad_edit(pad, path, kwargs, expect):